HTBHF-1670 rename do you have children
