+++ src/test/common/steps/pages.js
@@ -18,7 +18,7 @@ const ILiveInScotland = require('../page/i-live-in-scotland')
 const EmailAddress = require('../page/email-address')
 const UnsuccessfulApplication = require('../page/unsuccessful-application')
 const TermsAndConditions = require('../page/terms-and-conditions')
-const DoYouHaveChildrenThreeOrYounger = require('../page/do-you-have-children-three-or-younger')
+const DoYouHaveChildrenThreeOrYounger = require('../page/do-you-have-children')
 const SendCode = require('../page/send-code')
 const EnterChildrenDOB = require('../page/enter-children-dob')
 const EnterCode = require('../page/enter-code')
