--- src/web/routes/application/do-you-have-children-three-or-younger/index.js
@@ -1,5 +1,0 @@
-const { doYouHaveChildrenThreeOrYounger } = require('./do-you-have-children-three-or-younger')
-
-module.exports = {
-  doYouHaveChildrenThreeOrYounger
-}
