+++ src/web/routes/application/do-you-have-children/index.js
@@ -1,0 +1,5 @@
+const { doYouHaveChildrenThreeOrYounger } = require('./do-you-have-children')
+
+module.exports = {
+  doYouHaveChildrenThreeOrYounger
+}
