+++ src/web/routes/application/steps.js
@@ -7,7 +7,7 @@ const { phoneNumber } = require('./phone-number')
 const { doYouLiveInScotland } = require('./do-you-live-in-scotland')
 const { iLiveInScotland } = require('./i-live-in-scotland')
 const { emailAddress } = require('./email-address')
-const { doYouHaveChildrenThreeOrYounger } = require('./do-you-have-children-three-or-younger')
+const { doYouHaveChildrenThreeOrYounger } = require('./do-you-have-children')
 const { childrenDob } = require('./children-dob')
 const { sendCode } = require('./send-code')
 const { enterCode } = require('./enter-code')
